turn trusted users into a reusable component
